+++ src/routes/(secure)/tournaments/+page.server.ts
@@ -1,4 +1,4 @@
-import { redirect } from '@sveltejs/kit'
+import { fail, redirect } from '@sveltejs/kit'
 import { db } from '$lib/server/db';
 import { eq } from 'drizzle-orm';
 import { tournaments } from '$lib/server/db/schema';
@@ -15,7 +15,9 @@ export const actions: Actions = {
 	delete: async ( { request } ) => {
 		const formData = await request.formData();
 
-		const id = formData.get( 'id' );
+		const id = formData.get( 'id' ) as string;
+
+		if ( !id ) return fail( 400, { message: 'Missing id' } );
 
 		await db.delete( tournaments )
 			.where( eq( tournaments.id, id ) );
